+++ src/State/UserPasswordHasher.php
@@ -1,6 +1,4 @@
 <?php
-// api/src/State/UserPasswordHasher.php
-
 namespace App\State;
 
 use ApiPlatform\Metadata\Operation;
